TMP: dts: upstream: arm64: qcom: Add sm7150 community device DTS files
A lot of this is on the mailing list:
https://lore.kernel.org/linux-arm-msm/20250422-sm7150-upstream-v1-0-bf9a9081631d@jiaxyga.com/

The devicetrees for the two devices are not on the mailing list but will
be sent as soon as the SM7150 DTS landed.

We keep these around because they are sadly already in the community
category in postmarketOS without devicetrees upstream and want to get
rid of this as soon as possible.

Signed-off-by: Aelin <aelin@postmarketos.org>